pyproject.toml
[project]
name = "codec-finetuning"
version = "0.1.0"
requires-python = ">=3.12"
description = "Benchmark fine-tuning of neural speech codecs (Mimi, DualCodec, Kanade) on low-resource languages"
license = { text = "MIT" }
authors = [{ name = "codec-finetuning contributors" }]
readme = "README.md"
dependencies = [
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "moshi>=0.2.13",
    "dualcodec>=0.4.2",
    "kanade-tokenizer",
    "transformers>=4.46.0",
    "huggingface-hub>=0.25.0",
    "safetensors>=0.4.0",
    "einops>=0.8.0",
    "numpy>=1.26.0,<2.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "scipy>=1.12.0",
    "tqdm",
    "pyyaml",
]

[project.optional-dependencies]
train = [
    "accelerate>=1.2.0",
    "lightning>=2.5.0",
    "hydra-core>=1.3.0",
    "wandb>=0.19.0",
    "vocos>=0.1.0",
    "datasets>=3.0.0",
    "lion-pytorch>=0.2.0",
    "prodigyopt>=1.0",
    "schedulefree>=1.0",
    "adan-pytorch>=0.1.0",
]
eval = [
    "pesq==0.0.4",
    "pystoi>=0.4.0",
    "speechmos>=2.4.0",
    "jiwer>=3.0.0",
    "openai-whisper",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
torchaudio = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux'" }]
kanade-tokenizer = { git = "https://github.com/frothywater/kanade-tokenizer" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
